140925 by tomj on 2019/09/25 (pair: sweettea) ALB-3021 Merge change 140868 to Jasper 140868 by tomj on 2019/09/24 (pair: sweettea) ALB-3021 Merge change 140757 to linux-uds 140757 by tomj on 2019/09/16 (pair: sweettea) ALB-3021 Accept chapter index page headers in any endian order
.../deltaIndex.h
.../deltaIndex.c
Modify the initializeDeltaIndexPage method to accept an index page
header written in either endianness.  First read it in little endian
order.  If that fails, we then try big endian order.
Create the swapDeltaIndexPageEndianness method so we can test indices
written using the other endian order.

140926 by tomj on 2019/09/25 (pair: sweettea)
UDSUP-8 Merge change 140898 to Jasper
140898 by tomj on 2019/09/24 (pair: msakai)
UDSUP-8 Specify a UdsConfiguration when loading an existing index.

The reason to do this change is that it matches exactly what VDO is
able to do, and better matches what VDO wants UDS to do.  With the
UdsConfiguration, UDS can verify that the nonce of the existing index
matches the expectations of VDO.

.../uds.h
Add the expected UdsConfiguration as a required argument to the
udsLoadLocalIndex and udsRebuildLocalIndex methods.

.../udsMain.c
Add the expected UdsConfiguration as a required argument to the
udsLoadLocalIndex and udsRebuildLocalIndex methods, so that a
non-NULL UdsConfiguration is also a required argument to the
makeIndexSession method.
Store the UdsConfiguration in the new IndexSession as soon as we
have an IndexSession.  Then we can just pass the IndexSession
down to the lower level methods, and do not need to also pass
the UdsConfiguration down to those same methods.
When loading an existing index, verify that the expected
UdsConfiguration matches the configuration in the index.

.../indexLayout.h
.../indexLayout.c
When we are making an IndexLayout, we now always get the expected
UdsConfiguration.  Fail quickly if the size of the block device is
insufficient for the configuration.
Convert the readIndexConfig method into the verifyIndexConfig method.
We used to just read the UdsConfiguration from the index.  Now we
read it and compare it to make sure it is exactly matches the
expected UdsConfiguration.  This comparison is especially interesting
for the nonce, but it is useful to also ensure that the rest of the
configuration matches exactly.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>