141072 by tomj on 2019/10/14 (pair: sweettea) UDSUP-8 Merge change 141040 to Jasper 141040 by tomj on 2019/10/09 (pair: sweettea) UDSUP-8 Refactor opening an index and add runtime parameters.
.../uds.h
.../udsMain.c
Create a single udsOpenIndex method, replacing the udsCreateLocalIndex,
udsLoadLocalIndex and udsRebuildLocalIndex methods.
Add a typedef named UdsOpenIndexType, which is used for the first
to parameter to the udsOpenIndex method.
Add the struct uds_parameters for holding run time parameters, which is
used for one of the parameters to udsOpenIndex.

141073 by tomj on 2019/10/14 (pair: sweettea)
Merge change 141045 to Jasper
141045 by tomj on 2019/10/10 (pair: sweettea)
Change an index page map warning message.

.../indexPageMap.h
.../indexPageMap.c
Do not print a warning message when we change from chapter 0 to chapter
other than 1.  This case is normal and happens when we recreate the
index page map from scratch during a volume replay.  This change was
suggested during discussion of ALB-3022.
Delete the unused dumpIndexPageMap method.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>