141126 by tomj on 2019/10/17 (pair: sweettea) UDSUP-8 Merge change 141098 to Jasper 141098 by tomj on 2019/10/15 (pair: sweettea) UDSUP-8 Implement the checkpoint_frequency runtime parameter.
And stop using the checkpoint frequency from the UdsConfiguration.
Note that since we started using the kernel UDS index for VDO, this
UdsConfiguration parameter has always been zero.  There has been no
code path that allows the user to make any other choice.

.../uds.h
.../udsMain.c
Remove the XXX comment about checkpoint_frequency not being implemented
yet.
Delete the udsConfigurationGetCheckpointFrequency and
udsConfigurationGetCheckpointFrequency methods.
Just initialize the unused checkpoint frequency member of the
UdsConfiguration to the default value, which is zero.
Convert UDS_CHUNK_NAME_SIZE from a #define to an enum.

.../indexInternals.c
Set up the new index to using the checkpoint frequency from the
uds_parameters.

.../indexConfig.h
.../indexConfig.c
Remove the copy of the checkpoint frequency from the index
configuration.

.../config.h
.../config.c
Add an XXX note about the checkpoint frequency in the UdsConfiguration
not being used.  It is less trouble to just stop using the value than
to actually modify the structure, which would require a version
number change.
Remove usage of the checkpoint frequency from the UdsConfiguration.

141131 by tomj on 2019/10/17 (pair: sweettea)
UDSUP-8 Fix call to method deleted by merge 141126.

.../volumeGeometry.h
.../volumeGeometry.c
Add the indexConfigToUdsParameters method, and use it to set the
checkpoint frequence of the UDS index.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sweettea <sweettea@redhat.com>