141303 by tomj on 2019/10/29 (pair: sclafani) UDSUP-6 Merge change 141241 to Jasper 141241 by tomj on 2019/10/25 (pair: jwiele) UDSUP-6 Use correct non-deprecated interfaces to get the kernel time.
.../timeUtils.h
.../timeUtils.c
Replace the implementation of the currentTime method in kernel mode.
The former implementation called the getnstimeofday method, which has
been deprecated.  Obey the clock argument, choosing either the
get_time_ns method for monotonic time, or the get_time_real_ns
method for realtime.  Make the kernel method be static inline and
count on gcc removing the conditional and just calling the correct
method.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: jwiele <jwiele@redhat.com>