141763 by jwiele on 2019/12/12 (pair: sweettea) Merge 41742: Make FileUtils compatible with python2 and python3. 141742 by jwiele on 2019/12/11 (pair: jshimkus,sclafani,sweettea) Make FileUtils compatible with python2 and python3.
In python3 an iterator needs to define the __next__() method,
which will be called by the built-in next() function. In
python2, an iterator needs to define the next() method which
will be called directly as an object method. Defining
__next__() with next() as an alias satisfies both
versions.

With those methods defined, the object acts as a proper
iterator in both python2 and python3, so the __iter__() method
can return the object itself instead of delegating to the
contained python file object.

See:
https://portingguide.readthedocs.io/en/latest/iterators.html#new-iteration-protocol-next

.../FileUtils.py
Define both next() and __next__() methods to satisfy python2
and python3 iterator requirements. Define the __iter__()
method to return the object itself, since it is now a proper
iterator.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: jshimkus <jshimkus@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>