141995 by corwin on 2020/01/07 (pair: sclafani ) Merged 138027 to linux-vdo. 138027 by bgurney on 2019/02/13 (pair: raeburn) Remove read cache options from Ansible playbook templates. [VDO-4425]
[BZ 1663259]

VDO 6.2 no longer contains the "readCache" or "readCacheSize" options,
therefore the template playbooks for the Ansible VDO module should
no longer specify these options, to avoid suggesting that they
still exist.

If an Ansible playbook does specify the readCache or readCacheSize
options, vdomgmnt will throw the appropriate error to indicate
that the options are not recognized; the administrator can then
act on this error message, and remove the readCache or readCacheSize
options.

.../test_vdocreate_alloptions.yml
Remove the readCache and readCacheSize options.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: bgurney <bgurney@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>