142003 by corwin on 2020/01/07 (pair: sclafani ) Merged 138104 to linux-vdo. 138104 by joe on 2019/02/25 (pair: raeburn) Fewer errors when removing a vdo with a missing device (BZ 1535476).
This addresses an additional complaint in the reopened BZ about
excessive errors in the logs when we try and stop a vdo with its
backing device missing. There are no changes to the behavior of
the vdo command as seen by the user and the existing test
still applies.

.../VDOService.py
Move the check for a non-existent backing device to the
beginning of the remove function, before we attempt to stop.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>