142140 by bjohnston on 2020/01/16 (pair: sclafani) [BZ 1713749] [VDO-4622] Merge change 140213 to linux-vdo. 140213 by bjohnston on 2019/07/29 (pair: sweettea) [BZ 1713749] [VDO-4622] Add ability to modify a vdo uuid.
.../VDOArgumentParser.py
Add a uuid parameter and add it to the modify command.

.../Defaults.py
Add a default for uuid and some checking of the value
if the parameter is passed into vdo modify.

.../VDOService.py
Return error if you try to set all VDO volumes
to a specific uuid.
Skip trying to set a uuid on a running VDO volume.

.../vdo.8
Add doc to match the new uuid parameter in modify.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>