142200 by bjohnston on 2020/01/22 (pair: joe) [BZ 1713749] [VDO-4622] Merge 140353 to linux-vdo. 140353 by bjohnston on 2019/08/13 (pair: sweettea) [BZ 1713749] [VDO-4622] Add additional uuid support.
.../VDOArgumentParser.py
.../vdo.8
Add uuid parameter to create.

.../VDOService.py
Add uuid as a field in VDOService. This is done in order
to access uuid in commands like create which don't pass
in arguments.
In create, if a uuid is passed in, check to see if it already
exists in an exsiting VDO volume. If it doesn't set the
volume's uuid after we format the device.
Add uuid checking to modify and make sure to handle errors when
dealing with multiple modified values.
Add checkForExistingUUID to do the checking. The VDOs checked
will be all running VDO volumes plus all offline volumes in
the current config file.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: joe <jochapma@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>