142228 by bjohnston on 2020/01/24 (pair: sclafani) [BZ 1737619] [VDO-4727] Merge change 140450 to linux-vdo. 140450 by bjohnston on 2019/08/20 (pair: sweettea) [BZ 1737619] [VDO-4727] Add new import command.
.../VDOArgumentParser.py
Add a new vdo command called import which allows you to move
a VDO volume stored on disk into VDO manager.
Add uuid setting to import.

.../VDOOperation.py
Add support for new import command and make sure both name
and device are required as parameters.

.../vdo.8
Update vdo manager doc to include the import command.

.../VDOService.py
Add in a new state for the import command and a names hash for
all states to be able to set appropriate error messages.

Refactor some of the code in create into separate functions so
it can be used in the import command as well since they are
similar in parts.

Add an importDevice method to do the work of importing an
existing VDO into the VDO manager config file. Use the
refactored check and validate calls here. Get all
needed import fields from the storage device, making sure
not to override the uuid value if it is passed in.

Redo getConfigFromVDO to get all the data from vdodumpconfig,
not just the vdo specific fields. This will allow us to import
all needed info from disk with one call to vdodumpconfig.

Add a new function called getVDOConfigFromVDO to get just the
vdo fields. Have previous calls to getConfigFromVDO now call
this new method.

Use the beginImport operation state to import. Have it act the
same as create in that if we run into an issue part way through
importing, we want to remove the device out of the config file.
Add a names hash so that we display the correct operation name
for each state in our message.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>