142339 by sclafani on 2020/02/05 (pair: sweettea) Merged changes 142141 and 142145 to linux-vdo [BZ 1659173] 142141 by sclafani on 2020/01/16 (pair: sweettea) Made VDO manager display output from vdoformat [BZ 1659173]
The RFE calls for output to be displayed when the volume is
formatted, but currently all output from vdoformat is ignored
when it succeeds.

.../VDOService.py
Printed and logged the output from vdoformat (if any) one
indentation level below the command itself. Currently it's only
the number of logical blocks, when defaulted.

142145 by sclafani on 2020/01/16 (pair: sweettea)
BZ 1659173 [RFE] display max size based on slab size after formatting

example output from vdoformat:

Logical blocks defaulted to 22681282 blocks.
The VDO volume can address 87 GB in 698 data slabs, each 128 MB.
It can grow to address at most 1 TB of physical storage in 8192 slabs.
If a larger maximum size might be needed, use bigger slabs.

The first line is not new, but it wasn't being passed through by
VDO Manager prior to change 142141.

.../vdoFormat.c
Made vdoFormat load the VDO immediately after formatting it to query
the configuration. This also serves as a quick verification.
Used the loaded VDO to calculate the total slab space and
the maximal slab space for that volume. Added a function
to display that information on stdout, as asked for by the RFE.
In the spirit of the RFE, also advised using larger slabs to allow
the volume to eventually grow.
Added a function to print large data size values in a readable way.

.../vdoConfig.h
.../vdoConfig.c
Removed the logical blocks return pointer from formatVDO()
and formatVDOWithNonce() since that information is now obtained
by loading the new volume.

.../vdoFormat.c
Reflected changes.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>