142353 by sclafani on 2020/02/06 (pair: raeburn) Merged change 141334 to linux-vdo [VDO-4800] 141334 by corwin on 2019/11/01 (pair: sclafani) Fixed race in block map tree [VDO-4800].
Unit tests which use DataVIOs to populate the block map and then
trigger a recovery can run afoul of the assertion in drainZoneTrees()
that there are no active lookups when draining. This is due to
finishLookup() launching the DataVIO before decrementing the counter.
This isn't a problem in the kernel because recoveries are only done
before the VDO comes online and hence there can be no active DataVIOs.

.../blockMapTree.c
Modified finishLookup() to decrement the activeLookups counter before
calling the DataVIO's callback.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: raeburn <raeburn@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>