142377 by corwin on 2020/02/11 (pair: sclafani ) Merged 141158 to linux-vdo. 141158 by corwin on 2019/10/22 (pair: sclafani) Fixed infinite loop in action manager [VDO-4805].
If an action manager has a default scheduler, and the condition which
triggers the scheduler to attempt to schedule is true, but the manager
is in a state where the default action can't run, an infinite loop
results where each failure of the default action to run schedules the
action again.

.../actionManager.c
Modified finishActionCallback() to only attempt to run the scheduler
when the manager is in normal operation.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>