142383 by bjohnston on 2020/02/11 (pair: sweettea) [VDO-4818] [BZ1767491] Merge change 141328 to linux-vdo. 141328 by bjohnston on 2019/10/31 (pair: sclafani) [VDO-4818] [BZ1767491] Import works with no active dm devices.
We noticed an issue with device mapper where on machines that
had no active device mapper devices the output of a call to
dmsetup table --target vdo would return "No devices found"
instead of an empty string which is what was returned on
machines with active device mapper devices that were filtered
out by the --target parameter. This change makes it so we
handle both cases.

.../VDOService.py
Add checking for "No devices found" output from dmsetup
table.

Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>