142398 by corwin on 2020/02/12 (pair: sclafani ) Merged 141401 to linux-vdo. 141401 by corwin on 2019/11/12 (pair: msakai) Fixed premature slab drain notifications [VDO-4800] [BZ 1766358].
When draining a slab, the slab journal is told to drain. When it has
drained, it then tells the RefCounts to drain. When the RefCounts have
drained, they notify the slab which causes the slab to finish the
drain. The actual AdminState which informs both the journal and
RefCounts that they are being drained lives in the Slab and is shared
by both components. Unfortunately, if the last outstanding I/O from the
RefCounts happens to complete before the journal has finished draining,
the RefCounts will inform the slab that it has drained, and the drain
notification will be sent. This race was found by inspection and is
analogous to the block map zone race which caused VDO-4800. Although we
have seen no evidence of this issue causing problems in either the
kernel or unit tests, such issues can not be ruled out.

This change fixes the issue by modifying the drain complete check to be
done in a manner which checks that both components are drained before
sending the notification.

.../slab.h
.../slab.c
Added checkIfSlabDrained() which finishes the drain in a slab when both
the journal and the reference counts are drained. Modified
initiateSlabAction() to kick off a drain of both the journal and
reference counts when initiating a drain of the slab. Since we now
check both components explicitly, we don't need to wait to start the
reference counts drain until the journal is done (especially since
that didn't work correctly as spelled anyway). Doing so also ensures
that reference count or slab summary writes triggering slab journal
reaping can't happen after the slab has finished draining.

.../slabJournal.h
.../slabJournal.c
Renamed checkForDrainComplete() to isSlabJournalActive(), exposed it,
and modified it to merely return a boolean indicating whether the
journal is active (i.e. not done draining) rather than draining the
reference counts.
Replaced all calls to checkForDrainComplete() with calls to
checkIfSlabDrained().
Removed the call to checkForDrainComplete() from drainSlabJournal()
since this call is now redundant with the call in
initiateSlabAction().

.../refCounts.h
.../refCounts.c
Renamed checkForDrainComplete() to areRefCountsActive(), exposed it,
and modified it to merely return a boolean indicating whether the
reference counts are active (i.e. not done draining) rather than
directly finishing the drain.
Replaced all calls to checkForDrainComplete() with calls to
checkIfSlabDrained().
Removed the call to checkForDrainComplete() from drainRefCounts() since
this call is now redundant with the call in initiateSlabAction().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>